+++ packages/mistralai-gcp/RELEASES.md
@@ -49,3 +49,13 @@ Based on:
 - [typescript v1.1.0] packages/mistralai-gcp
 ### Releases
 - [NPM v1.1.0] https://www.npmjs.com/package/@mistralai/mistralai-gcp/v/1.1.0 - packages/mistralai-gcp
+
+## 2024-11-08 13:01:40
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.434.7 (2.452.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.3.0] packages/mistralai-gcp
+### Releases
+- [NPM v1.3.0] https://www.npmjs.com/package/@mistralai/mistralai-gcp/v/1.3.0 - packages/mistralai-gcp